sdk/pyproject.toml
[project]
name = "pixie-sdk"
version = "0.1.0"
description = "Local SDK server for Pixie AI Evals — manages datasets, renders labeling UI, and proxies to the remote Pixie server."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web server
    "fastapi>=0.115",
    "strawberry-graphql[fastapi]>=0.262",
    "uvicorn[standard]>=0.34",
    # Database
    "aiosqlite>=0.21",
    # File ingestion
    "polars>=1.0",
    "python-magic>=0.4",
    "pyarrow>=18",
    "genson>=1.3",
    # Templating
    "jinja2>=3.1",
    # Embedding
    "openai>=1.60",
    # Remote server client (ariadne-codegen output)
    "httpx>=0.28",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "mypy>=1.0",
    "ruff>=0.9",
    "ariadne-codegen>=0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py311"
